Add inline step option
Add an inline option to steps. Defaults to true.

Inline steps are all run in a single job execution. Steps that are not
inline will be run in a separate execution from surrounding steps.

This allows you to execute a long running step separately which is
useful to ensure that its progress is saved before it runs.